+++ packages/ui/src/components/index.ts
@@ -39,6 +39,7 @@ export {
   InputAdornment,
   Snackbar,
   Radio,
+  Select,
 } from '@mui/material';
 
 export type {
@@ -80,6 +81,7 @@ export type {
   SlideProps,
   GridProps,
   SnackbarProps,
+  SelectProps,
 } from '@mui/material';
 
 export * from './IconButton';
@@ -97,4 +99,5 @@ export * from './CopyButton';
 export * from './OtpInput';
 export * from './Typography';
 export * from './Banner';
+export * from './AmountInput';
 export * from './AddressDropdown';