+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ### vNext
 
-- 
+- [DAV-3724] Fix a console error when user closes auth popup
 
 ### v1.10.0
 
-- [DAV-3702] Fix skipping tour  
+- [DAV-3702] Fix skipping tour
 - [DAV-3674] Add collections to product tour
 - [DAV-3580] Display Cere mainnet address on UI