+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### vNext
 
 - [DAV-3664] Add Native Cere token balance to UI
+- [DAV-3724] Fix a console error when user closes auth popup
 
 ### v1.10.0